Plugin project structure
For the fundamental project structure, you need to use one of the
Gradle DSLs. Therefore, it is the same for all plugin implementation
approaches. Here we use the Kotlin DSL. You can also use the Groovy DSL
by removing the '.kts' extension from Gradle files. The syntax used
in the files is the same.