Register the Plugins
If you use the Kotlin DSL or the Groovy DSL, registering a plugin
only requires you to create a new script file in the 'src' folder.
The name of the file is the plugin ID. For the class-based approaches
you need some more boilerplate: Create a class that implements the
'Plugin' interface and assign an ID in the build file.