Compose plugins
Plugins usually apply other plugins and configure them. If you use
the Kotlin DSL or Groovy DSL, this is done by ID in the 'plugins {}'
block. In the class-based approaches, you call the 'plugins.apply()'
method. Here you can use the class of the other plugin which gives you
navigation in the IDE. But you may also use the plugin IDs here.