Configure a core plugin by Extension
Ideally, plugins provide Extensions to configure them. The Java plugins
that are delivered with Gradle provide the 'java { }' extension. As
an example, we use it to set the Java version to 17.
In the Kotlin DSL and the Groovy DSL, an extension is accessed directly
by its name. In class-baed approaches, we need to retrieve it via
'extensions.getByType()' and provide the implementation type of the
extension there.