Configure a task directly and add a dependency
Sometimes, tasks need to be configured directly. We do that on the
example of the 'test' task. Also, sometimes you want to add a dependency
to every project and hence want to do that in a plugin as well.
In the Kotlind DSL or Groovy DSL, tasks can be addressed by their name
through the 'tasks' container. Note that in Groovy DSL you should use
'tasks.named()' for task configuration avoidance. In the class-based
approaches, you also should use 'tasks.named()' and in addition you
need to provide the implementation type of the task (which is
<Test> in the example).
Dependencies are added through the 'dependencies' container which is
quite similar in all approaches.