Configure a comunity plugin by Extension
Similar to core plugins, also community plugins (or your own plugins)
may provide extensions to configure them.
In the Kotlin DSL and the Groovy DSL, an extension is accessed directly
by its name. In class-baed approaches, we need to retrieve it via
'extensions.getByType()' and provide the implementation type of the
extension there.